@@ -1,30 +1,31 @@
 #pragma once
 
+#include "YangContext.hpp"
+#include <exception>
+#include <libyang/libyang.h>
+#include <sstream>
 #include <stdexcept>
 #include <string>
-#include <exception>
-#include <vector>
-#include <sstream>
 #include <unordered_set>
-#include "YangContext.hpp"
-#include <libyang/libyang.h>
+#include <vector>
 
 namespace yang {
 
-class YangError : public std::exception {
-public:
-    YangError() {
-    }
+  class YangError : public std::exception {
+  public:
+    YangError() {}
 
     // what() uses the captured ly_err_item message if available
-    const char* what() const noexcept override {
-        static std::string buf;
-        buf = std::string("[No YANG ctx Available]: ") + getError();
-        return buf.c_str();
+    const char *what() const noexcept override {
+      static std::string buf;
+      buf = std::string("[No YANG ctx Available]: ") + getError();
+      return buf.c_str();
     }
 
     // Accessors for ly_err_item fields (lowercase)
-    virtual std::string getError() const { return std::string(ly_last_logmsg()); }
+    virtual std::string getError() const {
+      return std::string(ly_last_logmsg());
+    }
 
     // Two path accessors as requested
     virtual std::string getDataPath() const { return std::string(); }
@@ -38,97 +39,109 @@ public:
 
     virtual std::string getExpr() const { return std::string(); }
 
-protected:
-    const struct ly_err_item* error_ = nullptr; // only class property allowed
-    // Capture libyang error using provided context pointer; sets error_ and message_ and returns textual message if available.
-    const ly_err_item* captureError(const YangContext* ctx) {
-        return ly_err_last(ctx->raw());
+  protected:
+    const struct ly_err_item *error_ = nullptr; // only class property allowed
+    // Capture libyang error using provided context pointer; sets error_ and
+    // message_ and returns textual message if available.
+    const ly_err_item *captureError(const YangContext *ctx) {
+      return ly_err_last(ctx->raw());
     }
-};
+  };
 
-// YangDataError: constructed with a YangContext and calls CaptureError(&ctx)
-class YangDataError : public YangError {
-public:
-
-    explicit YangDataError(const YangContext& ctx) : ctx_(&ctx) {
-        error_ = captureError(&ctx);
+  // YangDataError: constructed with a YangContext and calls CaptureError(&ctx)
+  class YangDataError : public YangError {
+  public:
+    explicit YangDataError(const YangContext &ctx) : ctx_(&ctx) {
+      error_ = captureError(&ctx);
     }
 
     // Provide richer overrides that read from the captured ly_err_item chain
-    const char* what() const noexcept override {
-        if (what_buf_.empty()) {
-            std::string mod = getModuleName();
-            std::string err = getError();
-            std::string dpath = getDataPath();
-            std::string spath = getSchemaPath();
-            int line = getLine();
-            std::string apptag = getAppTag();
-            std::string expr = getExpr();
+    const char *what() const noexcept override {
+      if (what_buf_.empty()) {
+        std::string mod = getModuleName();
+        std::string err = getError();
+        std::string dpath = getDataPath();
+        std::string spath = getSchemaPath();
+        int line = getLine();
+        std::string apptag = getAppTag();
+        std::string expr = getExpr();
 
-            std::ostringstream ss;
-            if (!mod.empty()) ss << "[" << mod << "] ";
-            if (!err.empty()) ss << err;
-            if (!dpath.empty()) ss << "\nData path: " << dpath;
-            if (!spath.empty()) ss << "\nSchema path: " << spath;
-            if (line >= 0) ss << "\nLine: " << std::to_string(line);
-            if (!apptag.empty()) ss << "\nApp tag: " << apptag;
-            if (!expr.empty()) ss << "\nExpr: " << expr;
+        std::ostringstream ss;
+        if (!mod.empty())
+          ss << "[" << mod << "] ";
+        if (!err.empty())
+          ss << err;
+        if (!dpath.empty())
+          ss << "\nData path: " << dpath;
+        if (!spath.empty())
+          ss << "\nSchema path: " << spath;
+        if (line >= 0)
+          ss << "\nLine: " << std::to_string(line);
+        if (!apptag.empty())
+          ss << "\nApp tag: " << apptag;
+        if (!expr.empty())
+          ss << "\nExpr: " << expr;
 
-            // Append loaded modules from the context (if available)
-            if (ctx_ && ctx_->raw()) {
-                ss << "\nXXX this might be problematic: Loaded modules:";
-                uint32_t idx = 0;
-                const struct lys_module *m = nullptr;
-                while ((m = ly_ctx_get_module_iter(ctx_->raw(), &idx)) != NULL) {
-                    if (m->name) {
-                        ss << "\n - " << m->name;
-                    }
-                    ++idx;
-                }
+        // Append loaded modules from the context (if available)
+        if (ctx_ && ctx_->raw()) {
+          ss << "\nXXX this might be problematic: Loaded modules:";
+          uint32_t idx = 0;
+          const struct lys_module *m = nullptr;
+          while ((m = ly_ctx_get_module_iter(ctx_->raw(), &idx)) != NULL) {
+            if (m->name) {
+              ss << "\n - " << m->name;
             }
-
-            what_buf_ = ss.str();
+            ++idx;
+          }
         }
-        return what_buf_.c_str();
+
+        what_buf_ = ss.str();
+      }
+      return what_buf_.c_str();
     }
 
-private:
+  private:
     mutable std::string what_buf_;
 
     std::string getError() const override {
-        if (error_ && error_->msg) return std::string(error_->msg);
-        return YangError::getError();
+      if (error_ && error_->msg)
+        return std::string(error_->msg);
+      return YangError::getError();
     }
 
     std::string getDataPath() const override {
-        if (error_ && error_->data_path) return std::string(error_->data_path);
-        return YangError::getDataPath();
+      if (error_ && error_->data_path)
+        return std::string(error_->data_path);
+      return YangError::getDataPath();
     }
 
     std::string getSchemaPath() const override {
-        if (error_ && error_->schema_path) return std::string(error_->schema_path);
-        return YangError::getSchemaPath();
+      if (error_ && error_->schema_path)
+        return std::string(error_->schema_path);
+      return YangError::getSchemaPath();
     }
 
     int getLine() const override {
-        if (error_) return static_cast<int>(error_->line);
-        return YangError::getLine();
+      if (error_)
+        return static_cast<int>(error_->line);
+      return YangError::getLine();
     }
 
     std::string getAppTag() const override {
-        if (error_ && error_->apptag) return std::string(error_->apptag);
-        return YangError::getAppTag();
+      if (error_ && error_->apptag)
+        return std::string(error_->apptag);
+      return YangError::getAppTag();
     }
 
-    YangDataError(const YangDataError&) = default;
-    YangDataError& operator=(const YangDataError&) = default;
-    const YangContext* ctx_ = nullptr;
-};
+    YangDataError(const YangDataError &) = default;
+    YangDataError &operator=(const YangDataError &) = default;
+    const YangContext *ctx_ = nullptr;
+  };
 
-// Exception thrown when module iterator cannot continue (no more modules).
-class ModuleIteratorStopError : public std::runtime_error {
-public:
+  // Exception thrown when module iterator cannot continue (no more modules).
+  class ModuleIteratorStopError : public std::runtime_error {
+  public:
     ModuleIteratorStopError() : std::runtime_error(nullptr) {}
-};
+  };
 
 } // namespace yang
