@@ -1,255 +1,258 @@
 #pragma once
 
-#include <memory>
-#include <initializer_list>
 #include "YangContext.hpp"
-#include <vector>
+#include <initializer_list>
+#include <memory>
 #include <string>
-#include <utility>
 #include <tuple>
+#include <utility>
+#include <vector>
 
 namespace yang {
 
-class YangContext;
+  class YangContext;
 
-class Yang {
-public:
-    //Accept a variable-length list of YangContextOption values (default empty)
-    static std::shared_ptr<YangContext> getContext(std::initializer_list<YangContextOption> opts = {});
+  class Yang {
+  public:
+    // Accept a variable-length list of YangContextOption values (default empty)
+    static std::shared_ptr<YangContext>
+    getContext(std::initializer_list<YangContextOption> opts = {});
 
     // Return a fully-initialized default context: adds standard search paths
     // and loads a canonical set of modules. This initializes once and is
     // equivalent to calling `GetContext()` and then performing the runtime
     // search-path/module loading that the library typically requires.
     static std::shared_ptr<YangContext> getDefaultContext();
-};
+  };
 
-// Default module list with explicit revisions used by GetDefaultContext().
-// Now includes a third parameter (features pointer) currently set to nullptr.
-inline const std::vector<std::tuple<std::string,std::string,const char**>> kDefaultModules = {
-    // {"default","2025-06-18", nullptr},
-    // {"ietf-yang-schema-mount","2019-01-14", nullptr},
-    // {"ietf-datastores","2018-02-14", nullptr},
-    // {"ietf-yang-structure-ext","2020-06-17", nullptr},
-    // {"ietf-inet-types","2013-07-15", nullptr},
-    // {"ietf-yang-types","2013-07-15", nullptr},
-    // XXX shouldnt even load this I dont think.... Omit loading ietf-yang-library by default; tests provide any required yang-library
-    // {"ietf-yang-library","2019-01-04", nullptr},
-    // {"yang","2025-01-29", nullptr},
-    // {"ietf-yang-metadata","2016-08-05", nullptr},
-    // {"iana-bfd-types","2021-10-21", nullptr},
-    // {"iana-bgp-l2-encaps","2022-09-20", nullptr},
-    // {"iana-crypt-hash","2014-08-06", nullptr},
-    // {"iana-dots-signal-channel","2021-09-02", nullptr},
-    // {"iana-hardware","2018-03-13", nullptr},
-    // {"iana-msd-types","2025-01-10", nullptr},
-    // {"iana-pseudowire-types","2022-09-20", nullptr},
-    // {"iana-tls-profile","2025-04-18", nullptr},
-    // {"iana-tunnel-type","2019-11-16", nullptr},
-    // {"iana-bfd-types","2021-10-21", nullptr },
-    // {"iana-bgp-l2-encaps","2022-09-20", nullptr },
-    // {"iana-crypt-hash","2014-08-06", nullptr },
-    // {"iana-dns-class-rr-type","2025-05-20", nullptr },
-    // {"iana-dots-signal-channel","2021-09-02", nullptr },
-    // {"iana-hardware","2018-03-13", nullptr },
-    // {"iana-if-type","2021-06-21", nullptr },
-    // {"iana-if-type","2023-01-26", nullptr },
-    // {"iana-msd-types","2025-01-10", nullptr },
-    // {"iana-pseudowire-types","2022-09-20", nullptr },
-    // {"iana-routing-types","2025-02-18", nullptr },
-    // {"iana-ssh-encryption-algs","2024-10-16", nullptr },
-    // {"iana-ssh-key-exchange-algs","2024-10-16", nullptr },
-    // {"iana-ssh-mac-algs","2024-10-16", nullptr },
-    // {"iana-ssh-public-key-algs","2024-10-16", nullptr },
-    // {"iana-tls-cipher-suite-algs","2024-10-16", nullptr },
-    // {"iana-tls-profile","2025-04-18", nullptr },
-    // {"iana-tunnel-type","2021-04-23", nullptr },
-    // {"ietf-access-control-list","2019-03-04", nullptr},
-    // {"ietf-acldns","2019-01-28", nullptr},
-    // {"ietf-acl-tls","2025-04-18", nullptr},
-    // {"ietf-alarms","2022-06-06", nullptr},
-    // {"ietf-alarms-x733","2019-09-11", nullptr},
-    // {"ietf-babel","2024-10-10", nullptr},
-    // {"ietf-bfd","2022-09-22", nullptr},
-    // {"ietf-bfd-ip-mh","2022-09-22", nullptr},
-    // {"ietf-bfd-ip-sh","2022-09-22", nullptr},
-    // {"ietf-bfd-lag","2022-09-22", nullptr},
-    // {"ietf-bfd-large","2025-04-04", nullptr},
-    // {"ietf-bfd-mpls","2022-09-22", nullptr},
-    // {"ietf-bfd-types","2022-09-22", nullptr},
-    // {"ietf-bfd-unsolicited","2023-08-31", nullptr},
-    // {"ietf-complex-types","2011-03-15", nullptr},
-    // {"ietf-connectionless-oam","2019-04-16", nullptr},
-    // {"ietf-connectionless-oam-methods","2019-04-16", nullptr},
-    // {"ietf-connection-oriented-oam","2019-04-16", nullptr},
-    // {"ietf-crypto-types","2024-10-10", nullptr},
-    // {"ietf-dc-fabric-topology","2019-02-25", nullptr},
-    // {"ietf-dc-fabric-topology-state","2019-02-25", nullptr},
-    // {"ietf-dc-fabric-types","2019-02-25", nullptr},
-    // {"ietf-detnet","2024-10-28", nullptr},
-    // {"ietf-dhcpv6-client","2022-06-20", nullptr},
-    // {"ietf-dhcpv6-common","2022-06-20", nullptr},
-    // {"ietf-dhcpv6-relay","2022-06-20", nullptr},
-    // {"ietf-dhcpv6-server","2022-06-20", nullptr},
-    // {"ietf-dots-call-home","2021-12-09", nullptr},
-    // {"ietf-dots-data-channel","2020-05-28", nullptr},
-    // {"ietf-dots-mapping","2022-06-20", nullptr},
-    // {"ietf-dots-robust-trans","2023-02-28", nullptr},
-    // {"ietf-dots-signal-channel","2021-09-02", nullptr},
-    // {"ietf-dots-telemetry","2022-06-20", nullptr},
-    // {"ietf-ethernet-segment","2022-09-20", nullptr},
-    // {"ietf-ethertypes","2019-03-04", nullptr},
-    // {"ietf-factory-default","2020-08-31", nullptr},
-    // {"ietf-foo","2016-03-20", nullptr},
-    // {"ietf-geo-location","2022-02-11", nullptr},
-    // {"ietf-hardware","2018-03-13", nullptr},
-    // {"ietf-hardware-state","2018-03-13", nullptr},
-    // {"ietf-i2nsf-ike","2021-07-14", nullptr},
-    // {"ietf-i2nsf-ikec","2021-07-14", nullptr},
-    // {"ietf-i2nsf-ikeless","2021-07-14", nullptr},
-    // {"ietf-i2rs-rib","2018-09-13", nullptr},
-    // {"ietf-igmp-mld","2019-11-01", nullptr},
-    // {"ietf-igmp-mld-proxy","2023-05-30", nullptr},
-    // {"ietf-igmp-mld-snooping","2022-01-31", nullptr},
-    // {"ietf-inet-types","2013-07-15", nullptr},
-    // {"ietf-interface-protection","2019-06-19", nullptr},
-    // {"ietf-interfaces","2018-02-20", nullptr},
-    // {"ietf-ioam","2024-08-27", nullptr},
-    // {"ietf-ip","2018-02-22", nullptr},
-    // {"ietf-ipfix-psamp","2017-01-18", nullptr},
-    // {"ietf-ipsec-iptfs","2023-01-31", nullptr},
-    // {"ietf-ipv4-unicast-routing","2018-03-13", nullptr},
-    // {"ietf-ipv6-router-advertisements","2018-03-13", nullptr},
-    // {"ietf-ipv6-unicast-routing","2018-03-13", nullptr},
-    // {"ietf-isis","2022-10-19", nullptr},
-    // {"ietf-isis-reverse-metric","2022-10-19", nullptr},
-    // {"ietf-key-chain","2017-06-15", nullptr},
-    // {"ietf-keystore","2024-10-10", nullptr},
-    // {"ietf-l2-topology","2020-11-15", nullptr},
-    // {"ietf-l2-topology-state","2020-11-15", nullptr},
-    // {"ietf-l2vpn-ntw","2022-09-20", nullptr},
-    // {"ietf-l2vpn-svc","2018-10-09", nullptr},
-    // {"ietf-l3-unicast-topology","2018-02-26", nullptr},
-    // {"ietf-l3-unicast-topology-state","2018-02-26", nullptr},
-    // {"ietf-l3vpn-ntw","2022-02-14", nullptr},
-    // {"ietf-l3vpn-svc","2018-01-19", nullptr},
-    // {"ietf-layer0-types","2021-08-13", nullptr},
-    // {"ietf-lime-time-types","2019-04-16", nullptr},
-    // {"ietf-lmap-common","2017-08-08", nullptr},
-    // {"ietf-lmap-control","2017-08-08", nullptr},
-    // {"ietf-lmap-report","2017-08-08", nullptr},
-    // {"ietf-logical-network-element","2019-01-25", nullptr},
-    // {"ietf-microwave-radio-link","2019-06-19", nullptr},
-    // {"ietf-microwave-topology","2024-09-30", nullptr},
-    // {"ietf-microwave-types","2019-06-19", nullptr},
-    // {"ietf-module-tags","2021-01-04", nullptr},
-    // {"ietf-module-tags-state","2021-01-04", nullptr},
-    // {"ietf-mpls","2020-12-18", nullptr},
-    // {"ietf-mpls-ldp","2022-03-14", nullptr},
-    // {"ietf-mpls-ldp-extended","2022-03-14", nullptr},
-    // {"ietf-mpls-msd","2025-01-10", nullptr},
-    // {"ietf-msdp","2020-10-31", nullptr},
-    // {"ietf-mud","2019-01-28", nullptr},
-    // {"ietf-mud-tls","2025-04-18", nullptr},
-    // {"ietf-mud-transparency","2023-10-10", nullptr},
-    // {"ietf-network","2018-02-26", nullptr},
-    // {"ietf-network-instance","2019-01-21", nullptr},
-    // {"ietf-network-state","2018-02-26", nullptr},
-    // {"ietf-network-topology","2018-02-26", nullptr},
-    // {"ietf-network-topology-state","2018-02-26", nullptr},
-    // {"ietf-network-vpn-pm","2023-03-20", nullptr},
-    // {"ietf-nmda-compare","2021-12-10", nullptr},
-    // {"ietf-notification-capabilities","2022-02-17", nullptr},
-    // {"ietf-ntp","2022-07-05", nullptr},
-    // {"ietf-origin","2018-02-14", nullptr},
-    // {"ietf-ospf","2022-10-19", nullptr},
-    // {"ietf-ospfv3-extended-lsa","2024-06-07", nullptr},
-    // {"ietf-packet-fields","2019-03-04", nullptr},
-    // {"ietf-pim-base","2022-10-19", nullptr},
-    // {"ietf-pim-bidir","2022-10-19", nullptr},
-    // {"ietf-pim-dm","2022-10-19", nullptr},
-    // {"ietf-pim-rp","2022-10-19", nullptr},
-    // {"ietf-pim-sm","2022-10-19", nullptr},
-    // {"ietf-ptp","2019-05-07", nullptr},
-    // {"ietf-restconf","2017-01-26", nullptr},
-    // {"ietf-restconf-monitoring","2017-01-26", nullptr},
-    // {"ietf-restconf-subscribed-notifications","2019-11-17", nullptr},
-    // {"ietf-rib-extension","2023-11-20", nullptr},
-    // {"ietf-rift","2025-04-04", nullptr},
-    // {"ietf-rip","2020-02-20", nullptr},
-    // {"ietf-routing","2018-03-13", nullptr},
-    // {"ietf-routing-policy","2021-10-11", nullptr},
-    // {"ietf-routing-types","2017-12-04", nullptr},
-    // {"ietf-sap-ntw","2023-06-20", nullptr},
-    // {"ietf-schc","2023-03-01", nullptr},
-    // {"ietf-schc-compound-ack","2023-07-26", nullptr},
-    // {"ietf-segment-routing","2021-05-26", nullptr},
-    // {"ietf-segment-routing-common","2021-05-26", nullptr},
-    // {"ietf-segment-routing-mpls","2021-05-26", nullptr},
-    // {"ietf-service-assurance","2023-07-11", nullptr},
-    // {"ietf-service-assurance-device","2023-07-11", nullptr},
-    // {"ietf-service-assurance-interface","2023-07-11", nullptr},
-    // {"ietf-sid-file","2024-07-31", nullptr},
-    // {"ietf-snmp","2014-12-10", nullptr},
-    // {"ietf-snmp-common","2014-12-10", nullptr},
-    // {"ietf-snmp-community","2014-12-10", nullptr},
-    // {"ietf-snmp-engine","2014-12-10", nullptr},
-    // {"ietf-snmp-notification","2014-12-10", nullptr},
-    // {"ietf-snmp-proxy","2014-12-10", nullptr},
-    // {"ietf-snmp-ssh","2014-12-10", nullptr},
-    // {"ietf-snmp-target","2014-12-10", nullptr},
-    // {"ietf-snmp-tls","2014-12-10", nullptr},
-    // {"ietf-snmp-tsm","2014-12-10", nullptr},
-    // {"ietf-snmp-usm","2014-12-10", nullptr},
-    // {"ietf-snmp-vacm","2014-12-10", nullptr},
-    // {"ietf-softwire-br","2019-11-16", nullptr},
-    // {"ietf-softwire-ce","2019-11-16", nullptr},
-    // {"ietf-softwire-common","2019-11-16", nullptr},
-    // {"ietf-ssh-client","2024-10-10", nullptr},
-    // {"ietf-ssh-common","2024-10-10", nullptr},
-    // {"ietf-ssh-server","2024-10-10", nullptr},
-    // {"ietf-subscribed-notifications","2019-09-09", nullptr},
-    // {"ietf-syslog","2025-04-30", nullptr},
-    // {"ietf-system","2014-08-06", nullptr},
-    // {"ietf-system-capabilities","2022-02-17", nullptr},
-    // {"ietf-system-tacacs-plus","2021-08-05", nullptr},
-    // {"ietf-sztp-bootstrap-server","2019-04-30", nullptr},
-    // {"ietf-sztp-conveyed-info","2019-04-30", nullptr},
-    // {"ietf-sztp-csr","2024-10-10", nullptr},
-    // {"ietf-tcg-algs","2024-12-05", nullptr},
-    // {"ietf-tcp","2024-10-10", nullptr},
-    // {"ietf-tcp-client","2024-10-10", nullptr},
-    // {"ietf-tcp-common","2024-10-10", nullptr},
-    // {"ietf-tcp-server","2024-10-10", nullptr},
-    // {"ietf-template","2016-03-20", nullptr},
-    // {"ietf-te-packet-types","2020-06-10", nullptr},
-    // {"ietf-te-topology","2020-08-06", nullptr},
-    // {"ietf-te-topology-state","2020-08-06", nullptr},
-    // {"ietf-te-types","2020-06-10", nullptr},
-    // {"ietf-tls-client","2024-10-10", nullptr},
-    // {"ietf-tls-common","2024-10-10", nullptr},
-    // {"ietf-tls-server","2024-10-10", nullptr},
-    // {"ietf-tpm-remote-attestation","2024-12-05", nullptr},
-    // {"ietf-truststore","2024-10-10", nullptr},
-    // {"ietf-twamp","2021-11-17", nullptr},
-    // {"ietf-vn","2025-03-27", nullptr},
-    // {"ietf-voucher","2018-05-09", nullptr},
-    // {"ietf-voucher-request","2021-05-20", nullptr},
-    // {"ietf-vpn-common","2022-02-11", nullptr},
-    // {"ietf-vrrp","2018-03-13", nullptr},
-    // {"ietf-wson-topology","2021-08-13", nullptr},
-    // {"ietf-x509-cert-to-name","2014-12-10", nullptr},
-    // {"ietf-ztp-types","2024-10-10", nullptr}
-};
+  // Default module list with explicit revisions used by GetDefaultContext().
+  // Now includes a third parameter (features pointer) currently set to nullptr.
+  inline const std::vector<std::tuple<std::string, std::string, const char **>>
+      kDefaultModules = {
+          // {"default","2025-06-18", nullptr},
+          // {"ietf-yang-schema-mount","2019-01-14", nullptr},
+          // {"ietf-datastores","2018-02-14", nullptr},
+          // {"ietf-yang-structure-ext","2020-06-17", nullptr},
+          // {"ietf-inet-types","2013-07-15", nullptr},
+          // {"ietf-yang-types","2013-07-15", nullptr},
+          // XXX shouldnt even load this I dont think.... Omit loading
+          // ietf-yang-library by default; tests provide any required
+          // yang-library
+          // {"ietf-yang-library","2019-01-04", nullptr},
+          // {"yang","2025-01-29", nullptr},
+          // {"ietf-yang-metadata","2016-08-05", nullptr},
+          // {"iana-bfd-types","2021-10-21", nullptr},
+          // {"iana-bgp-l2-encaps","2022-09-20", nullptr},
+          // {"iana-crypt-hash","2014-08-06", nullptr},
+          // {"iana-dots-signal-channel","2021-09-02", nullptr},
+          // {"iana-hardware","2018-03-13", nullptr},
+          // {"iana-msd-types","2025-01-10", nullptr},
+          // {"iana-pseudowire-types","2022-09-20", nullptr},
+          // {"iana-tls-profile","2025-04-18", nullptr},
+          // {"iana-tunnel-type","2019-11-16", nullptr},
+          // {"iana-bfd-types","2021-10-21", nullptr },
+          // {"iana-bgp-l2-encaps","2022-09-20", nullptr },
+          // {"iana-crypt-hash","2014-08-06", nullptr },
+          // {"iana-dns-class-rr-type","2025-05-20", nullptr },
+          // {"iana-dots-signal-channel","2021-09-02", nullptr },
+          // {"iana-hardware","2018-03-13", nullptr },
+          // {"iana-if-type","2021-06-21", nullptr },
+          {"iana-if-type", "2023-01-26", nullptr},
+          // {"iana-msd-types","2025-01-10", nullptr },
+          // {"iana-pseudowire-types","2022-09-20", nullptr },
+          // {"iana-routing-types","2025-02-18", nullptr },
+          // {"iana-ssh-encryption-algs","2024-10-16", nullptr },
+          // {"iana-ssh-key-exchange-algs","2024-10-16", nullptr },
+          // {"iana-ssh-mac-algs","2024-10-16", nullptr },
+          // {"iana-ssh-public-key-algs","2024-10-16", nullptr },
+          // {"iana-tls-cipher-suite-algs","2024-10-16", nullptr },
+          // {"iana-tls-profile","2025-04-18", nullptr },
+          // {"iana-tunnel-type","2021-04-23", nullptr },
+          // {"ietf-access-control-list","2019-03-04", nullptr},
+          // {"ietf-acldns","2019-01-28", nullptr},
+          // {"ietf-acl-tls","2025-04-18", nullptr},
+          // {"ietf-alarms","2022-06-06", nullptr},
+          // {"ietf-alarms-x733","2019-09-11", nullptr},
+          // {"ietf-babel","2024-10-10", nullptr},
+          // {"ietf-bfd","2022-09-22", nullptr},
+          // {"ietf-bfd-ip-mh","2022-09-22", nullptr},
+          // {"ietf-bfd-ip-sh","2022-09-22", nullptr},
+          // {"ietf-bfd-lag","2022-09-22", nullptr},
+          // {"ietf-bfd-large","2025-04-04", nullptr},
+          // {"ietf-bfd-mpls","2022-09-22", nullptr},
+          // {"ietf-bfd-types","2022-09-22", nullptr},
+          // {"ietf-bfd-unsolicited","2023-08-31", nullptr},
+          // {"ietf-complex-types","2011-03-15", nullptr},
+          // {"ietf-connectionless-oam","2019-04-16", nullptr},
+          // {"ietf-connectionless-oam-methods","2019-04-16", nullptr},
+          // {"ietf-connection-oriented-oam","2019-04-16", nullptr},
+          // {"ietf-crypto-types","2024-10-10", nullptr},
+          // {"ietf-dc-fabric-topology","2019-02-25", nullptr},
+          // {"ietf-dc-fabric-topology-state","2019-02-25", nullptr},
+          // {"ietf-dc-fabric-types","2019-02-25", nullptr},
+          // {"ietf-detnet","2024-10-28", nullptr},
+          // {"ietf-dhcpv6-client","2022-06-20", nullptr},
+          // {"ietf-dhcpv6-common","2022-06-20", nullptr},
+          // {"ietf-dhcpv6-relay","2022-06-20", nullptr},
+          // {"ietf-dhcpv6-server","2022-06-20", nullptr},
+          // {"ietf-dots-call-home","2021-12-09", nullptr},
+          // {"ietf-dots-data-channel","2020-05-28", nullptr},
+          // {"ietf-dots-mapping","2022-06-20", nullptr},
+          // {"ietf-dots-robust-trans","2023-02-28", nullptr},
+          // {"ietf-dots-signal-channel","2021-09-02", nullptr},
+          // {"ietf-dots-telemetry","2022-06-20", nullptr},
+          // {"ietf-ethernet-segment","2022-09-20", nullptr},
+          // {"ietf-ethertypes","2019-03-04", nullptr},
+          // {"ietf-factory-default","2020-08-31", nullptr},
+          // {"ietf-foo","2016-03-20", nullptr},
+          // {"ietf-geo-location","2022-02-11", nullptr},
+          // {"ietf-hardware","2018-03-13", nullptr},
+          // {"ietf-hardware-state","2018-03-13", nullptr},
+          // {"ietf-i2nsf-ike","2021-07-14", nullptr},
+          // {"ietf-i2nsf-ikec","2021-07-14", nullptr},
+          // {"ietf-i2nsf-ikeless","2021-07-14", nullptr},
+          // {"ietf-i2rs-rib","2018-09-13", nullptr},
+          // {"ietf-igmp-mld","2019-11-01", nullptr},
+          // {"ietf-igmp-mld-proxy","2023-05-30", nullptr},
+          // {"ietf-igmp-mld-snooping","2022-01-31", nullptr},
+          // {"ietf-inet-types","2013-07-15", nullptr},
+          // {"ietf-interface-protection","2019-06-19", nullptr},
+          {"ietf-interfaces", "2018-02-20", nullptr},
+          // {"ietf-ioam","2024-08-27", nullptr},
+          {"ietf-ip", "2018-02-22", nullptr},
+          // {"ietf-ipfix-psamp","2017-01-18", nullptr},
+          // {"ietf-ipsec-iptfs","2023-01-31", nullptr},
+          // {"ietf-ipv4-unicast-routing","2018-03-13", nullptr},
+          // {"ietf-ipv6-router-advertisements","2018-03-13", nullptr},
+          // {"ietf-ipv6-unicast-routing","2018-03-13", nullptr},
+          // {"ietf-isis","2022-10-19", nullptr},
+          // {"ietf-isis-reverse-metric","2022-10-19", nullptr},
+          // {"ietf-key-chain","2017-06-15", nullptr},
+          // {"ietf-keystore","2024-10-10", nullptr},
+          // {"ietf-l2-topology","2020-11-15", nullptr},
+          // {"ietf-l2-topology-state","2020-11-15", nullptr},
+          // {"ietf-l2vpn-ntw","2022-09-20", nullptr},
+          // {"ietf-l2vpn-svc","2018-10-09", nullptr},
+          // {"ietf-l3-unicast-topology","2018-02-26", nullptr},
+          // {"ietf-l3-unicast-topology-state","2018-02-26", nullptr},
+          // {"ietf-l3vpn-ntw","2022-02-14", nullptr},
+          // {"ietf-l3vpn-svc","2018-01-19", nullptr},
+          // {"ietf-layer0-types","2021-08-13", nullptr},
+          // {"ietf-lime-time-types","2019-04-16", nullptr},
+          // {"ietf-lmap-common","2017-08-08", nullptr},
+          // {"ietf-lmap-control","2017-08-08", nullptr},
+          // {"ietf-lmap-report","2017-08-08", nullptr},
+          // {"ietf-logical-network-element","2019-01-25", nullptr},
+          // {"ietf-microwave-radio-link","2019-06-19", nullptr},
+          // {"ietf-microwave-topology","2024-09-30", nullptr},
+          // {"ietf-microwave-types","2019-06-19", nullptr},
+          // {"ietf-module-tags","2021-01-04", nullptr},
+          // {"ietf-module-tags-state","2021-01-04", nullptr},
+          // {"ietf-mpls","2020-12-18", nullptr},
+          // {"ietf-mpls-ldp","2022-03-14", nullptr},
+          // {"ietf-mpls-ldp-extended","2022-03-14", nullptr},
+          // {"ietf-mpls-msd","2025-01-10", nullptr},
+          // {"ietf-msdp","2020-10-31", nullptr},
+          // {"ietf-mud","2019-01-28", nullptr},
+          // {"ietf-mud-tls","2025-04-18", nullptr},
+          // {"ietf-mud-transparency","2023-10-10", nullptr},
+          // {"ietf-network","2018-02-26", nullptr},
+          // {"ietf-network-instance","2019-01-21", nullptr},
+          // {"ietf-network-state","2018-02-26", nullptr},
+          // {"ietf-network-topology","2018-02-26", nullptr},
+          // {"ietf-network-topology-state","2018-02-26", nullptr},
+          // {"ietf-network-vpn-pm","2023-03-20", nullptr},
+          // {"ietf-nmda-compare","2021-12-10", nullptr},
+          // {"ietf-notification-capabilities","2022-02-17", nullptr},
+          // {"ietf-ntp","2022-07-05", nullptr},
+          // {"ietf-origin","2018-02-14", nullptr},
+          // {"ietf-ospf","2022-10-19", nullptr},
+          // {"ietf-ospfv3-extended-lsa","2024-06-07", nullptr},
+          // {"ietf-packet-fields","2019-03-04", nullptr},
+          // {"ietf-pim-base","2022-10-19", nullptr},
+          // {"ietf-pim-bidir","2022-10-19", nullptr},
+          // {"ietf-pim-dm","2022-10-19", nullptr},
+          // {"ietf-pim-rp","2022-10-19", nullptr},
+          // {"ietf-pim-sm","2022-10-19", nullptr},
+          // {"ietf-ptp","2019-05-07", nullptr},
+          // {"ietf-restconf","2017-01-26", nullptr},
+          // {"ietf-restconf-monitoring","2017-01-26", nullptr},
+          // {"ietf-restconf-subscribed-notifications","2019-11-17", nullptr},
+          // {"ietf-rib-extension","2023-11-20", nullptr},
+          // {"ietf-rift","2025-04-04", nullptr},
+          // {"ietf-rip","2020-02-20", nullptr},
+          // {"ietf-routing","2018-03-13", nullptr},
+          // {"ietf-routing-policy","2021-10-11", nullptr},
+          // {"ietf-routing-types","2017-12-04", nullptr},
+          // {"ietf-sap-ntw","2023-06-20", nullptr},
+          // {"ietf-schc","2023-03-01", nullptr},
+          // {"ietf-schc-compound-ack","2023-07-26", nullptr},
+          // {"ietf-segment-routing","2021-05-26", nullptr},
+          // {"ietf-segment-routing-common","2021-05-26", nullptr},
+          // {"ietf-segment-routing-mpls","2021-05-26", nullptr},
+          // {"ietf-service-assurance","2023-07-11", nullptr},
+          // {"ietf-service-assurance-device","2023-07-11", nullptr},
+          // {"ietf-service-assurance-interface","2023-07-11", nullptr},
+          // {"ietf-sid-file","2024-07-31", nullptr},
+          // {"ietf-snmp","2014-12-10", nullptr},
+          // {"ietf-snmp-common","2014-12-10", nullptr},
+          // {"ietf-snmp-community","2014-12-10", nullptr},
+          // {"ietf-snmp-engine","2014-12-10", nullptr},
+          // {"ietf-snmp-notification","2014-12-10", nullptr},
+          // {"ietf-snmp-proxy","2014-12-10", nullptr},
+          // {"ietf-snmp-ssh","2014-12-10", nullptr},
+          // {"ietf-snmp-target","2014-12-10", nullptr},
+          // {"ietf-snmp-tls","2014-12-10", nullptr},
+          // {"ietf-snmp-tsm","2014-12-10", nullptr},
+          // {"ietf-snmp-usm","2014-12-10", nullptr},
+          // {"ietf-snmp-vacm","2014-12-10", nullptr},
+          // {"ietf-softwire-br","2019-11-16", nullptr},
+          // {"ietf-softwire-ce","2019-11-16", nullptr},
+          // {"ietf-softwire-common","2019-11-16", nullptr},
+          // {"ietf-ssh-client","2024-10-10", nullptr},
+          // {"ietf-ssh-common","2024-10-10", nullptr},
+          // {"ietf-ssh-server","2024-10-10", nullptr},
+          // {"ietf-subscribed-notifications","2019-09-09", nullptr},
+          // {"ietf-syslog","2025-04-30", nullptr},
+          // {"ietf-system","2014-08-06", nullptr},
+          // {"ietf-system-capabilities","2022-02-17", nullptr},
+          // {"ietf-system-tacacs-plus","2021-08-05", nullptr},
+          // {"ietf-sztp-bootstrap-server","2019-04-30", nullptr},
+          // {"ietf-sztp-conveyed-info","2019-04-30", nullptr},
+          // {"ietf-sztp-csr","2024-10-10", nullptr},
+          // {"ietf-tcg-algs","2024-12-05", nullptr},
+          // {"ietf-tcp","2024-10-10", nullptr},
+          // {"ietf-tcp-client","2024-10-10", nullptr},
+          // {"ietf-tcp-common","2024-10-10", nullptr},
+          // {"ietf-tcp-server","2024-10-10", nullptr},
+          // {"ietf-template","2016-03-20", nullptr},
+          // {"ietf-te-packet-types","2020-06-10", nullptr},
+          // {"ietf-te-topology","2020-08-06", nullptr},
+          // {"ietf-te-topology-state","2020-08-06", nullptr},
+          // {"ietf-te-types","2020-06-10", nullptr},
+          // {"ietf-tls-client","2024-10-10", nullptr},
+          // {"ietf-tls-common","2024-10-10", nullptr},
+          // {"ietf-tls-server","2024-10-10", nullptr},
+          // {"ietf-tpm-remote-attestation","2024-12-05", nullptr},
+          // {"ietf-truststore","2024-10-10", nullptr},
+          // {"ietf-twamp","2021-11-17", nullptr},
+          // {"ietf-vn","2025-03-27", nullptr},
+          // {"ietf-voucher","2018-05-09", nullptr},
+          // {"ietf-voucher-request","2021-05-20", nullptr},
+          // {"ietf-vpn-common","2022-02-11", nullptr},
+          // {"ietf-vrrp","2018-03-13", nullptr},
+          // {"ietf-wson-topology","2021-08-13", nullptr},
+          // {"ietf-x509-cert-to-name","2014-12-10", nullptr},
+          // {"ietf-ztp-types","2024-10-10", nullptr}
+  };
 
-// Default filesystem search paths used by GetDefaultContext().
-inline const std::vector<std::string> kDefaultSearchPaths = {
-    "/usr/local/share/yang/modules/libyang/",
-    "/usr/local/share/yang/modules/yang/standard/iana/",
-    "/usr/local/share/yang/modules/yang/standard/ietf/RFC/",
-    "/usr/share/yang/modules/libyang/",
-    "/usr/share/yang/modules/yang/standard/iana/",
-    "/usr/share/yang/modules/yang/standard/ietf/RFC/"
-};
+  // Default filesystem search paths used by GetDefaultContext().
+  inline const std::vector<std::string> kDefaultSearchPaths = {
+      "/usr/local/share/yang/modules/libyang/",
+      "/usr/local/share/yang/modules/yang/standard/iana/",
+      "/usr/local/share/yang/modules/yang/standard/ietf/RFC/",
+      "/usr/share/yang/modules/libyang/",
+      "/usr/share/yang/modules/yang/standard/iana/",
+      "/usr/share/yang/modules/yang/standard/ietf/RFC/"};
 
-} //namespace yang
+} // namespace yang
