@@ -1,15 +1,15 @@
 #pragma once
 
+#include "YangSchemaModule.hpp"
+#include <libyang/libyang.h>
 #include <string>
 #include <vector>
-#include <libyang/libyang.h>
-#include "YangSchemaModule.hpp"
 
 namespace yang {
-// Map libyang context flags (LY_CTX_*) to a scoped enum for callers.
-// Each enum member corresponds to a libyang LY_CTX_* macro (see libyang/context.h).
-// Brief descriptions are included here for convenience.
-enum class YangContextOptions : unsigned {
+  // Map libyang context flags (LY_CTX_*) to a scoped enum for callers.
+  // Each enum member corresponds to a libyang LY_CTX_* macro (see
+  // libyang/context.h). Brief descriptions are included here for convenience.
+  enum class YangContextOptions : unsigned {
     None = 0u,
 
     // LY_CTX_ALL_IMPLEMENTED (0x01)
@@ -18,12 +18,14 @@ enum class YangContextOptions : unsigned {
 
     // LY_CTX_REF_IMPLEMENTED (0x02)
     // Implement all imported modules "referenced" from an implemented module
-    // (implement modules referenced by when/must/defaults so referenced nodes work).
+    // (implement modules referenced by when/must/defaults so referenced nodes
+    // work).
     RefImplemented = LY_CTX_REF_IMPLEMENTED,
 
     // LY_CTX_NO_YANGLIBRARY (0x04)
-    // Do not internally implement ietf-yang-library module; ly_ctx_get_yanglib_data()
-    // will not work until the module is loaded manually.
+    // Do not internally implement ietf-yang-library module;
+    // ly_ctx_get_yanglib_data() will not work until the module is loaded
+    // manually.
     NoYanglibrary = LY_CTX_NO_YANGLIBRARY,
 
     // LY_CTX_DISABLE_SEARCHDIRS (0x08)
@@ -36,7 +38,8 @@ enum class YangContextOptions : unsigned {
     DisableSearchdirCwd = LY_CTX_DISABLE_SEARCHDIR_CWD,
 
     // LY_CTX_PREFER_SEARCHDIRS (0x20)
-    // When searching for schema files prefer configured searchdirs over callback.
+    // When searching for schema files prefer configured searchdirs over
+    // callback.
     PreferSearchdirs = LY_CTX_PREFER_SEARCHDIRS,
 
     // LY_CTX_EXPLICIT_COMPILE (0x80)
@@ -50,7 +53,8 @@ enum class YangContextOptions : unsigned {
     SetPrivParsed = LY_CTX_SET_PRIV_PARSED,
 
     // LY_CTX_COMPILE_OBSOLETE (0x0200)
-    // Include obsolete nodes in compiled schema so they can be created/validated.
+    // Include obsolete nodes in compiled schema so they can be
+    // created/validated.
     CompileObsolete = LY_CTX_COMPILE_OBSOLETE,
 
     // LY_CTX_LYB_HASHES (0x0400)
@@ -71,45 +75,51 @@ enum class YangContextOptions : unsigned {
     BuiltinPluginsOnly = LY_CTX_BUILTIN_PLUGINS_ONLY,
 
     // LY_CTX_STATIC_PLUGINS_ONLY (0x4000)
-    // Restrict loading to static (built-in) plugins and ignore external plugin dirs.
+    // Restrict loading to static (built-in) plugins and ignore external plugin
+    // dirs.
     StaticPluginsOnly = LY_CTX_STATIC_PLUGINS_ONLY,
 
     // LY_CTX_ENABLE_IMP_FEATURES (0x0100)
     // Enable all features of newly implemented imported modules by default.
     EnableImpFeatures = LY_CTX_ENABLE_IMP_FEATURES
-};
+  };
 
-using YangContextOption = YangContextOptions;
+  using YangContextOption = YangContextOptions;
 
-inline constexpr unsigned toFlags(YangContextOptions o) noexcept { return static_cast<unsigned>(o); }
+  inline constexpr unsigned toFlags(YangContextOptions o) noexcept {
+    return static_cast<unsigned>(o);
+  }
 
-class YangContext {
-public:
+  class YangContext {
+  public:
     YangContext(unsigned options = 0);
-    explicit YangContext(struct ly_ctx* raw_ctx) noexcept;
+    explicit YangContext(struct ly_ctx *raw_ctx) noexcept;
     ~YangContext();
 
-    YangContext(const YangContext&) = delete;
-    YangContext& operator=(const YangContext&) = delete;
+    YangContext(const YangContext &) = delete;
+    YangContext &operator=(const YangContext &) = delete;
 
-    void addSearchPath(const std::string& path);
-    struct lys_module* loadModuleInContext(const std::string& name,
-                                           const std::string& revision,
+    void addSearchPath(const std::string &path);
+    struct lys_module *loadModuleInContext(const std::string &name,
+                                           const std::string &revision,
                                            const char **features = nullptr);
-    const std::vector<std::string>& searchPaths() const noexcept { return search_paths_; }
+    const std::vector<std::string> &searchPaths() const noexcept {
+      return search_paths_;
+    }
 
-    struct ly_ctx* raw() const noexcept { return ctx_; }
+    struct ly_ctx *raw() const noexcept { return ctx_; }
 
     // (Removed) ModuleIterator helper - iteration is done directly via
     // libyang APIs in source to keep the header minimal.
 
-    // Find a loaded module by name; returns an empty YangSchemaModule if not found.
-    YangSchemaModule GetLoadedModuleByName(const std::string& name) const;
+    // Find a loaded module by name; returns an empty YangSchemaModule if not
+    // found.
+    YangSchemaModule GetLoadedModuleByName(const std::string &name) const;
 
-private:
-    struct ly_ctx* ctx_ = nullptr;
+  private:
+    struct ly_ctx *ctx_ = nullptr;
     std::vector<std::string> search_paths_;
     unsigned options_ = 0;
-};
+  };
 
 } // namespace yang
